backend/requirements.txt
# miniClaw Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5

# LangChain Community & Experimental
langchain-community>=0.0.10
langchain-experimental>=0.0.40

# LangGraph for Tree of Thoughts / PERV / Reflection StateGraph
langgraph>=1.0.10
langgraph-checkpoint-sqlite>=3.0.3

# LlamaIndex (RAG)
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.0
llama-index-retrievers-bm25>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-embeddings-ollama>=0.1.0
llama-index-embeddings-gemini>=0.1.0
llama-index-vector-stores-chroma>=0.1.0

# Vector Store
chromadb>=0.4.0

# Document Processing
pypdf>=3.0.0
python-docx>=1.0.0
docx2txt>=0.8
openpyxl>=3.1.0  # XLSX (Excel 2007+) file support
xlrd>=2.0.0  # XLS (Excel 97-2003) file support
sentence-transformers>=2.2.0

# Data Processing & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.0  # YAML parsing for skill frontmatter

# HTTP & Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
arxiv>=2.0.0  # arXiv paper search API (for arxiv-search skill)

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.0  # System monitoring (memory, CPU, process)

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing FastAPI

# Code Quality
ruff>=0.1.0
black>=23.12.0
mypy>=1.7.0
